% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_species.R
\name{all_species}
\alias{all_species}
\title{Common species name in a matrix form. This approach can result in
issues in identifying species. Needs to be improved in order to create
search groups for complex species.}
\usage{
all_species()
}
\description{
Common species name in a matrix form. This approach can result in
issues in identifying species. Needs to be improved in order to create
search groups for complex species.
}
\author{
Chantel Wetzel
}
