% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rank.R
\name{calculate_rank}
\alias{calculate_rank}
\title{Calculate the assessment prioritization rank by species based upon all the factors.}
\usage{
calculate_rank(
  fishing_mortality,
  commercial_importance,
  tribal_importance,
  recreational_importance,
  ecosystem,
  stock_status,
  assessment_frequency,
  constituent_demand,
  new_information,
  rebuilding
)
}
\arguments{
\item{fishing_mortality}{Output from summarize_fishing_mortality function}

\item{commercial_importance}{Output from summarize_revenue for commercial fisheries}

\item{tribal_importance}{Output from summarize_revenue for tribal fisheries}

\item{recreational_importance}{Output from summarize_rec_importance function}

\item{ecosystem}{Output from summarize_ecosytem function}

\item{stock_status}{Output from summarize_stock_status function}

\item{assessment_frequency}{Output from summarize_frequency function}

\item{constituent_demand}{Output from summarize_const_demand function}

\item{new_information}{Output from summarize_new_information function}

\item{rebuilding}{Output from summarize_rebuilding function}
}
\description{
Calculate the assessment prioritization rank by species based upon all the factors.
}
\author{
Chantel Wetzel
}
