% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ties.R
\name{calculate_ties}
\alias{calculate_ties}
\title{Calculate tiess within ranks.}
\usage{
calculate_ties(data, column_name = "Rank")
}
\arguments{
\item{data}{Dataframe}

\item{column_name}{Column name to determine the ties for}
}
\description{
Calculate tiess within ranks.
}
\author{
Chantel Wetzel
}
