% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_files.R
\name{clean_model_files}
\alias{clean_model_files}
\title{Function to delete extraneous Stock Synthesis files.
This function only deletes extra Stock Synthesis files to reduce
the number of files for each assessment. Not necessary to run to
read or summarize model results.}
\usage{
clean_model_files(loc = "model_files/")
}
\arguments{
\item{loc}{file location where folders needing cleaned up are}
}
\description{
Function to delete extraneous Stock Synthesis files.
This function only deletes extra Stock Synthesis files to reduce
the number of files for each assessment. Not necessary to run to
read or summarize model results.
}
\author{
Chantel Wetzel
}
