% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calender.R
\name{create_calender}
\alias{create_calender}
\title{Create a calender}
\usage{
create_calender(dir = here::here("doc/figs"), year = 2025)
}
\arguments{
\item{dir}{Directory}

\item{year}{Year to create a calender for.}
}
\description{
Create a calender
}
\author{
Chantel Wetzel
}
