% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_area_gemm.R
\name{filter_area_gemm}
\alias{filter_area_gemm}
\title{Filter the GEMM data for select sectors}
\usage{
filter_area_gemm(data)
}
\arguments{
\item{data}{Dataframe created by nwfscSurvey::pull_gemm}
}
\description{
Filter the GEMM data for select sectors
}
\author{
Chantel Wetzel
}
