% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gemm.R
\name{filter_gemm}
\alias{filter_gemm}
\title{Filter the GEMM data}
\usage{
filter_gemm(
  data,
  sector = c("California Recreational", "Oregon Recreational", "Washington Recreational")
)
}
\arguments{
\item{data}{Dataframe created by nwfscSurvey::pull_gemm}

\item{sector}{List of sectors to retain}
}
\description{
Filter the GEMM data
}
\author{
Chantel Wetzel
}
