% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_revenue.R
\name{filter_revenue}
\alias{filter_revenue}
\title{Filter out and identify tribal and non-tribal records}
\usage{
filter_revenue(data, type = NULL)
}
\arguments{
\item{data}{PacFIN CompFT data table}

\item{type}{Default NULL. To filter and mark tribal records type = "tribal"}
}
\description{
Filter out and identify tribal and non-tribal records
}
\author{
Chantel Wetzel
}
