% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_years.R
\name{filter_years}
\alias{filter_years}
\title{Filter data by year}
\usage{
filter_years(data, years)
}
\arguments{
\item{data}{Dataframe}

\item{years}{Vector of years to retain}
}
\description{
Filter data by year
}
\author{
Chantel Wetzel
}
