% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_all.R
\name{format_all}
\alias{format_all}
\title{Do all final formatting}
\usage{
format_all(
  x,
  man_groups = readr::read_csv(here::here("data-raw", "species_management_groups.csv"))
)
}
\arguments{
\item{x}{A factor data frame that contains species names to be renamed.}

\item{col_names}{The species column name within the factor data frame}
}
\value{
The factor data frame with the species renamed
}
\description{
Do all final formatting
}
\author{
Chantel Wetzel
}
