% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_species_names.R
\name{format_species_names}
\alias{format_species_names}
\title{Standardize species names}
\usage{
format_species_names(x)
}
\arguments{
\item{x}{A factor data frame that contains species names to be renamed.}
}
\value{
The factor data frame with the species renamed
}
\description{
Function to format species names to add cryptic species names is
missing and to apply the upper case for the first letter for each
species name.
}
\examples{
\dontrun{ mort_adj <- format_species_names(x = mortality) }

}
\author{
Chantel Wetzel
}
