% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_files_to_sap.R
\name{move_files_to_sap}
\alias{move_files_to_sap}
\title{Copy and move csv file by factor to the shiny repo}
\usage{
move_files_to_sap(
  dir = here::here("data-processed"),
  move_to = "C:/Users/Chantel.Wetzel/Documents/gitHub/prioritization/wcgfishSAP/tables"
)
}
\arguments{
\item{dir}{Directory of saved factor file}

\item{move_to}{Directory of the shiny app}
}
\description{
Copy and move csv file by factor to the shiny repo
}
\author{
Chantel Wetzel
}
