% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_const_demand.R
\name{summarize_const_demand}
\alias{summarize_const_demand}
\title{Calculations used for the "Const Demand" tab for assessment prioritization.}
\usage{
summarize_const_demand(
  revenue_data,
  rec_importance_data,
  fishing_mortality,
  future_spex,
  species = species
)
}
\arguments{
\item{revenue_data}{Filtered revenue by year (filter_years) from PacFIN that
includes both commercial and tribal revenue (data-raw/pacfin_revenue.csv).}

\item{rec_importance_data}{Data object created by summarize_rec_importance function}

\item{fishing_mortality}{Data object created by summarize_fishing_mortality function}

\item{future_spex}{Data objected created from the data-raw/GMT008-harvest specifications_alt2-2025.csv}

\item{species}{Data object read from the data folder called "species_names.csv" that includes
all the species to include in this analysis..}
}
\description{
The values used in the constituent demand tab are primarily scored qualitatively.
This function will provide the state, gear, and sector (commercial vs. recreational)
differences across the states and coastwide which then can be qualitatively
used to input modifiers.
}
\author{
Chantel Wetzel
}
