% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ecosystem.R
\name{summarize_ecosystem}
\alias{summarize_ecosystem}
\title{Calculate ecosystem rank based upon Atlantis/Ecosim output}
\usage{
summarize_ecosystem(ecosystem_data)
}
\arguments{
\item{ecosystem_data}{A csv file with ecosystem scores by species based on Atlantis/Ecosim}
}
\description{
Calculate ecosystem rank based upon Atlantis/Ecosim output
}
\author{
Chantel Wetzel
}
