% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_future_spex.R
\name{summarize_future_spex}
\alias{summarize_future_spex}
\title{Function to compare future ACLs to existing fishing mortality averages.
This function uses output from the summarize_fishing_mortality and the summarize_frequency
functions and hence should be only run after these two functions. The manage_file
is downloaded from PacFIN APEX future harvest specifications
table (GMT008). The draft harvest specification ACLs should correspond to
the current year plus two (e.g., 2022 + 2 = 2024 ACLs). The fishing_mort_file
is the csv file created by the summarize_fishing_mortality function where
the average fishing mortality across recent years is used to compare future
ACLs against. The freq_file is a csv file created by the summarize_frequency function where
the year of the last assessment will be found to be appending in this output file.}
\usage{
summarize_future_spex(manage_file, fishing_mort_file, freq_file, species_file)
}
\arguments{
\item{manage_file}{A csv file with OFLs and ACLs from the draft harvest
specifications table in PacFIN APEX reporting online.}

\item{fishing_mort_file}{A csv file created by the summarize_fishing_mortality function.}

\item{freq_file}{A csv file created by the summarize_frequency function.}

\item{species_file}{A csv file in the data folder called "species_names.csv" to summarize across.}
}
\description{
Function to compare future ACLs to existing fishing mortality averages.
This function uses output from the summarize_fishing_mortality and the summarize_frequency
functions and hence should be only run after these two functions. The manage_file
is downloaded from PacFIN APEX future harvest specifications
table (GMT008). The draft harvest specification ACLs should correspond to
the current year plus two (e.g., 2022 + 2 = 2024 ACLs). The fishing_mort_file
is the csv file created by the summarize_fishing_mortality function where
the average fishing mortality across recent years is used to compare future
ACLs against. The freq_file is a csv file created by the summarize_frequency function where
the year of the last assessment will be found to be appending in this output file.
}
\examples{

Example call including all species:
summarize_future_spex(
   manage_file <- "GMT008-harvest specifications-2024.csv",
   fishing_mort_file <- "fishing_mortality.csv",
   freq_file <- "assessment_frequency.csv",
   species_file <-  "species_names.csv"
)

}
\author{
Chantel Wetzel
}
