% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_results.R
\name{summarize_model_results}
\alias{summarize_model_results}
\title{Function that will read output from all models in
the model_loc ("model_files") folder. The unfished
spawning biomass, final spawning biomass, sigma R,
assessment year, and mean age. The function then takes
these results and calculates a weighted depletion and
mean age for stocks with area based assessments. These
results are then added to the previous cycle's stock status
and assessment frequency sheets in the "data" folder. The
updated stock status and assessment frequency csv files are
then saved to the tables folder.}
\usage{
summarize_model_results(
  abundance_file,
  frequency_file,
  species_file,
  model_loc = "model_files",
  years
)
}
\arguments{
\item{abundance_file}{A csv file containing the existing abundance by species that the
newly assessed species abundance estimates will be added to or old abundance values replaced. The
csv file to be read is found in the data folder.}

\item{frequency_file}{A csv file from the previous assessment prioritization assessment
frequency tab. The csv file to be read is found in the data folder.}

\item{model_loc}{Folder name to look for model files. The default is "model_files" in the
assessment prioritization github.}

\item{years}{Vector of specific years to calculate the mean age of the catches by species.}
}
\description{
Function that will read output from all models in
the model_loc ("model_files") folder. The unfished
spawning biomass, final spawning biomass, sigma R,
assessment year, and mean age. The function then takes
these results and calculates a weighted depletion and
mean age for stocks with area based assessments. These
results are then added to the previous cycle's stock status
and assessment frequency sheets in the "data" folder. The
updated stock status and assessment frequency csv files are
then saved to the tables folder.
}
\examples{

summarize_model_results(
		abundance_file <- "abundance_previous_cycle.csv",
		frequency_file <- "assessment_frequency_last_cycle.csv",
		species_file <- "species_names.csv",
		model_loc <- "model_files",
		years <- 2000:2020
)

}
\author{
Chantel Wetzel
}
