% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_new_information.R
\name{summarize_new_information}
\alias{summarize_new_information}
\title{Calculate score based upon new information and research}
\usage{
summarize_new_information(species, survey_data, assess_year, new_research)
}
\arguments{
\item{species}{CSV file in the data folder called "species_names.csv" that includes
all the species to include in this analysis.}

\item{survey_data}{Data frame of WCGBTS bds data from data-processed/all_nwfsc_survey_new_information.csv}

\item{assess_year}{R data object with the assessment year by species from the
data-processed/assess_year_ssc_rec.csv}

\item{new_research}{R data object that contains new research to be considered
scoring. New research by species contained in data-raw/new_research.csv.}
}
\description{
Calculate score based upon new information and research
}
\author{
Chantel Wetzel
}
