% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rebuilding.R
\name{summarize_rebuilding}
\alias{summarize_rebuilding}
\title{Calculate score based upon rebuilding}
\usage{
summarize_rebuilding(species, overfished_data, stock_status, assessment_year)
}
\arguments{
\item{species}{CSV file in the data folder called "species_names.csv" that includes
all the species to include in this analysis.}

\item{overfished_data}{Data frame of overfished species from data-raw/overfished_species.csv}

\item{stock_status}{Data frame created by the summarize_stock_status function}

\item{assessment_year}{The year for which species to be assessed are being selected based
upon the prioritization process}
}
\description{
Calculate score based upon rebuilding
}
\author{
Chantel Wetzel
}
