% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_revenue.R
\name{summarize_revenue}
\alias{summarize_revenue}
\title{Summarize and format commercial revenue data for insert
into the assessment prioritizaiton. Data are pulled from
PacFIN filtering for only "P" Council records and removing
all tribal and research landing revenue estimates.  The data file
includes the following columns from PacFIN:
AGENCY_CODE, COUNCIL_CODE, PACFIN_SPECIES_CODE, PACFIN_SPECIES_COMMON_NAME,
NOMINAL_TO_ACTUAL_PACFIN_SPECIES_CODE, PACFIN_YEAR, FLEET_CODE, ROUND_WEIGHT_MTONS,
AFI_EXVESSEL_REVENUE. The AFI_EXVESSEL_REVENUE column adjusts for inflation and the
NOMINAL_TO_ACTUAL_PACFIN_SPECIES_CODE includes both nominal and species specific
records. Exvessel revenue is averaged over select years by species and dollar values
are output in the 1,000s.}
\usage{
summarize_revenue(file_name, species_file, years, max_exp = 0.18)
}
\arguments{
\item{file_name}{CSV file pulled from PacFIN with commerical revenue
data}

\item{species_file}{A csv file will all species name to summarize}

\item{years}{A vector of years to calculate the average revenue across (e.g. 2016:2020)}

\item{max_exp}{A numerical value to assume as the exponent to create a reasonable
spread across rankings by species.}
}
\description{
Summarize and format commercial revenue data for insert
into the assessment prioritizaiton. Data are pulled from
PacFIN filtering for only "P" Council records and removing
all tribal and research landing revenue estimates.  The data file
includes the following columns from PacFIN:
AGENCY_CODE, COUNCIL_CODE, PACFIN_SPECIES_CODE, PACFIN_SPECIES_COMMON_NAME,
NOMINAL_TO_ACTUAL_PACFIN_SPECIES_CODE, PACFIN_YEAR, FLEET_CODE, ROUND_WEIGHT_MTONS,
AFI_EXVESSEL_REVENUE. The AFI_EXVESSEL_REVENUE column adjusts for inflation and the
NOMINAL_TO_ACTUAL_PACFIN_SPECIES_CODE includes both nominal and species specific
records. Exvessel revenue is averaged over select years by species and dollar values
are output in the 1,000s.
}
\examples{
summarize_revenue(
		file_name <- "revenue_summarized_12052021.csv"
		species_file <-  "species_names.csv"
		years <- 2016:2020
)

summarize_revenue(
		file_name <- "tribal_revenue_12062021.csv"
		species_file <-  "species_names.csv"
		years <- 2016:2020
)


}
\author{
Chantel Wetzel
}
