% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Combine management group to tables based on species names}
\usage{
format_table(x, man_groups)
}
\arguments{
\item{x}{A factor dataframe}

\item{man_groups}{A dataframe containing the species management groups}
}
\value{
Joined dataframe with edited column names
}
\description{
Function to join the given table with the table that has the species management
groups. Replaces underscores in column names with spaces.
}
\author{
Chantel Wetzel
}
