% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fishing_mortality.R
\name{summarize_fishing_mortality}
\alias{summarize_fishing_mortality}
\title{Comparison between recent average mortality, OFLs, and ACLs.}
\usage{
summarize_fishing_mortality(
  gemm_mortality,
  harvest_spex,
  species = species,
  manage_quants = c("OVERFISHING_LIMIT", "ANNUAL_CATCH_LIMIT")
)
}
\arguments{
\item{gemm_mortality}{Groundfish expanded multiyear mortality by species.
R object created by \code{\link[nwfscSurvey:pull_gemm]{nwfscSurvey::pull_gemm()}}.}

\item{harvest_spex}{R data objected filetered by year using \code{\link[=filter_years]{filter_years()}}
containing historical groundfish harvest specifications OFLs, ABCs, and ACLs across
years for West Coast groundfish. The data should be downloaded from PacFIN
APEX report table 15 (or provided by Jason Edwards at PacStates) and should
be stored in the data-raw folder (example: data-raw/GMT015-final specifications-2015 - 2023.csv)}

\item{species}{R data object that contains a list of species names to calculate
assessment prioritization.  The csv file with the list of species names should be
stored in the data-raw folder ("species_names.csv")}

\item{manage_quants}{The names of the management quantities in the harvest_spex data
object that defines the OFL and ACL. This vector is used to grep those columns.
This allows to easily shift between the ABC and the ACL if needed.}
}
\description{
The official recent average mortality should
come from the GEMM rather than a pull from PacFIN.
The manage_file is a file containing management quantities
provided in a special pull by Rob Ames from the GMT016 table.
This special pull of all management quantities by species across years
had some issues that had to be fixed by hand:
\enumerate{
\item Many rockfish species were missing "Rockfish" in their names and
I had to correct them by hand due to grep issues when only the first
name were used (e.g., China changed to China Rockfish).
\item The file does not contain OFLs or ACLs for sablefish, longspine
thornyhead, and shortspine thornyhead.
\item Quantities (OFLs,..) were duplicated for blue rockfish and rougheye/blackspotted
rockfish and had to be corrected by hand.
}
}
\details{
2023 Notes of GMT 015 correction
\enumerate{
\item blue rockfish in Oregon - hand deleted the blue/deacon/black complex rows
\item cowcod - hand deleted the south of 4010 rows and add in the ACLs by area
\item lingcod - hand deleted WA-OR and 42-4010 rows
\item starry and kelp - add rockfish to each
}

2023 future spex fixes
\enumerate{
\item delete s4010 cowcod
}
}
\author{
Chantel Wetzel
}
