% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_frequency.R
\name{summarize_frequency}
\alias{summarize_frequency}
\title{Function to create the full assessment frequency tab.}
\usage{
summarize_frequency(
  abundance,
  ecosystem,
  commercial,
  tribal,
  recreational,
  assessment_year = 2027
)
}
\arguments{
\item{abundance}{R object of suggested assessment frequency based upon biology. A csv file from
the previous assessment prioritization assessment
frequency tab. The csv file to be read is found in the data folder:
data-processessed/previous_cycle/abundance_processed.csv}

\item{ecosystem}{R data object created by \code{\link[=summarize_ecosystem]{summarize_ecosystem()}} for tribal catch
data.}

\item{commercial}{R data object created by \code{\link[=summarize_revenue]{summarize_revenue()}} for commercial catch
data.}

\item{tribal}{R data object created by \code{\link[=summarize_revenue]{summarize_revenue()}} for tribal catch
data.}

\item{recreational}{R data object created by \code{\link[=summarize_rec_importance]{summarize_rec_importance()}} for tribal catch
data.}

\item{assessment_year}{A numerical value of the current year the assessment prioritization
is being conducted. This value is compared to the last year assessed values to provide species
rotation in the ranking based on time since the last assessment.}
}
\description{
Function to create the full assessment frequency tab.
}
\author{
Chantel Wetzel
}
