% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_new_information.R
\name{summarize_new_information}
\alias{summarize_new_information}
\title{Calculate score based upon new information and research}
\usage{
summarize_new_information(
  species = species,
  survey_data,
  bio_params,
  new_research
)
}
\arguments{
\item{species}{CSV file in the data folder called "species_names.csv" that includes
all the species to include in this analysis.}

\item{survey_data}{Data frame of WCGBTS bds data from data-processed/all_nwfsc_survey_new_information.csv}

\item{bio_params}{Data frame that includes the last year a species was assessed from
"data-processed/species_sigmaR_catage.csv"}

\item{new_research}{Data frame of new research by species contained in data-raw/new_research.csv}
}
\description{
Calculate score based upon new information and research
}
\author{
Chantel Wetzel
}
