% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rec_importance.R
\name{summarize_rec_importance}
\alias{summarize_rec_importance}
\title{Calculate recreational importance}
\usage{
summarize_rec_importance(rec_catch, species, rec_importance, assess_year)
}
\arguments{
\item{rec_catch}{R data object created by filter_gemm function with only recreational catches
by state.}

\item{species}{R data object that contains a list of species names to calculate
assessment prioritization.  The csv file with the list of species names should be
stored in the data-raw folder ("species_names.csv")}

\item{rec_importance}{R data object with the species importance scoring by state.
A CSV file with the state-specific species importance scores}

\item{assess_year}{R data object with the assessment year by species from the
data-processed/assess_year_ssc_rec.csv.}
}
\description{
Summarize and format recreational catch data to be used along with recreational
importance scores to calculate the 'pseudo revenue' by species for the recreational fishery.
The function currently uses an existing csv file with previously calculated recreational
importance to access existing relative weights for each species and state. In the future
this should be modified in the future to use a stand-alone file containing the recreational
species weights by state that should be saved in the "data" folder.
}
\examples{
\dontrun{
summarize_rec_importance(
		file_name <- "CTE002-2016---2020.csv",
		years <- 2016:2020,
		species_file <-  "species_names.csv"
)
}


}
\author{
Chantel Wetzel
}
