% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rec_importance.R
\name{summarize_rec_importance}
\alias{summarize_rec_importance}
\title{Summarize and format recreational catch data to be used along with recreational
importance scores to calculate the 'pseudo revenue' by species for the recreational fishery.
The function currently uses an existing csv file with previously calculated recreational
importance to access existing relative weights for each species and state. In the future
this should be modified in the future to use a stand-alone file containing the recreational
species weights by state that should be saved in the "data" folder.}
\usage{
summarize_rec_importance(file_name, species_file, years, max_exp = 0.18)
}
\arguments{
\item{file_name}{A csv file pulled from RecFIN with recreational catches.
data. Found in the data folder in the assessment prioritization github repo.}

\item{species_file}{A csv including all species to calculate values for.}

\item{years}{A vector of years to calculate the average recreational catch
across.}

\item{max_exp}{A numerical value to apply as an exponent to the coastwide
pseudo revenue score. Default value of 0.18.}
}
\description{
Summarize and format recreational catch data to be used along with recreational
importance scores to calculate the 'pseudo revenue' by species for the recreational fishery.
The function currently uses an existing csv file with previously calculated recreational
importance to access existing relative weights for each species and state. In the future
this should be modified in the future to use a stand-alone file containing the recreational
species weights by state that should be saved in the "data" folder.
}
\examples{

summarize_rec_importance(
		file_name <- "CTE002-2016---2020.csv",
		years <- 2016:2020,
		species_file <-  "species_names.csv"
)


}
\author{
Chantel Wetzel
}
