% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_revenue.R
\name{summarize_revenue}
\alias{summarize_revenue}
\title{Calculate importance by revenue}
\usage{
summarize_revenue(revenue, species, tribal_score = NULL, frequency)
}
\arguments{
\item{revenue}{CSV file pulled from PacFIN with commerical revenue
data}

\item{species}{Data object read from the data folder called "species_names.csv" that includes
all the species to include in this analysis.}

\item{tribal_score}{CSV with the tribal importance score by species}

\item{frequency}{Suggested assessment frequency based upon biology. A csv file from
the previous assessment prioritization assessment
frequency tab. The csv file to be read is found in the data folder:
data-raw/species_sigmaR_catage_main.csv}
}
\description{
Summarize and format commercial revenue data for insert
into the assessment prioritizaiton. Data are pulled from
PacFIN filtering for only "P" Council records and removing
all tribal and research landing revenue estimates.  The data file
includes the following columns from PacFIN:
AGENCY_CODE, COUNCIL_CODE, PACFIN_SPECIES_CODE, PACFIN_SPECIES_COMMON_NAME,
NOMINAL_TO_ACTUAL_PACFIN_SPECIES_CODE, PACFIN_YEAR, FLEET_CODE, ROUND_WEIGHT_MTONS,
AFI_EXVESSEL_REVENUE. The AFI_EXVESSEL_REVENUE column adjusts for inflation and the
NOMINAL_TO_ACTUAL_PACFIN_SPECIES_CODE includes both nominal and species specific
records. Exvessel revenue is averaged over select years by species and dollar values
are output in the 1,000s.
}
\author{
Chantel Wetzel
}
